api-principal/requirements.txt
asgiref==3.8.1
Django
django-cors-headers
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.8.0
djongo==1.3.7
dnspython==2.7.0
mongoengine==0.29.1
pymongo==3.11.4
pytz==2024.2
setuptools==75.6.0
sqlparse==0.2.4
tzdata==2024.2
pillow==11.2.1
mysqlclient
requests
drf-yasg
